Functions

Detailed explanations or additional notes about all functions.

Measurement functions

BlueTangle.sample_outcomes — Function

sample_outcomes(state::SparseVector, shots::Int) -> Vector

Sample outcomes from a quantum state vector based on the probability distribution.

state: A sparse quantum state vector.

shots: Number of samples to be taken.

Returns a vector of sampled outcomes.

BlueTangle.get_probabilities_from_sample — Function

get_probabilities_from_sample(sample_outcomes::Vector, N::Int) -> (Vector, Vector)

Convert a sample of outcomes into probabilities.

sample_outcomes: A vector of sampled outcomes.

N: Number of qubits.

Returns a tuple of vectors: the first vector contains outcomes, and the second vector contains corresponding probabilities.

BlueTangle.get_expect — Function

Alias for density matrix:

get_expect(state::SparseVector, op::QuantumOps) -> Float64
get_expect(state::SparseVector, op_str::String, qubit::Int) -> Float64
get_expect(state::SparseVector, matrix::SparseMatrixCSC) -> Float64
get_expect(state::SparseVector, op_str::String) -> Vector{Float64}

Alias for state vector:

get_expect(rho::SparseMatrixCSC, op::QuantumOps) -> Float64
get_expect(rho::SparseMatrixCSC, op_str::String, qubit::Int) -> Float64
get_expect(rho::SparseMatrixCSC, matrix::SparseMatrixCSC) -> Float64
get_expect(rho::SparseMatrixCSC, op_str::String) -> Vector{Float64}

Calculate the expectation value for quantum states or density matrices given an operator. This function has several forms depending on the input parameters:

get_expect(state::SparseVector, op::QuantumOps): Computes the expectation value for a quantum state vector with a given operator.

get_expect(rho::SparseMatrixCSC, op::QuantumOps): Computes the expectation value for a density matrix with a given operator.

get_expect(state::SparseVector, op_str::String, qubit::Int): Computes the expectation value for a specific qubit in a quantum state vector with an operator specified as a string.

get_expect(rho::SparseMatrixCSC, op_str::String, qubit::Int): Computes the expectation value for a specific qubit in a density matrix with an operator specified as a string.

get_expect(state::SparseVector, op_str::String): Computes the expectation values for all qubits in a quantum state vector given an operator as a string.

get_expect(rho::SparseMatrixCSC, op_str::String): Computes the expectation values for all qubits in a density matrix given an operator as a string.

get_expect(state::SparseVector, matrix::SparseMatrixCSC): Computes the expectation value using a sparse matrix representation of an operator for a state vector.

get_expect(rho::SparseMatrixCSC, matrix::SparseMatrixCSC): Computes the expectation value using a sparse matrix representation of an operator for a density matrix.

Arguments

state::SparseVector: The quantum state vector.

rho::SparseMatrixCSC: The density matrix.

op::QuantumOps: The quantum operator.

op_str::String: The string representation of the operator.

qubit::Int: The specific qubit index.

matrix::SparseMatrixCSC: The sparse matrix representation of the operator.

Returns

The expectation value as a Float64 or a vector of Float64 for multiple qubits.

BlueTangle.get_expect_from_measurement — Function

Alias:

get_expect_from_measurement(m::Measurement, qubit::Int) -> Float64
get_expect_from_measurement(m::Measurement) -> Vector

Calculate the expectation value from a measurement.

m: A Measurement object.

qubit: (Optional) Specific qubit for which to calculate the expectation value.

Returns the expectation value(s).

BlueTangle.get_corr — Function

Alias:

get_corr(state::SparseVector, list_of_operators::String, qubits_applied::Vector) -> Float64
get_corr(rho::SparseMatrixCSC, list_of_operators::String, qubits_applied::Vector) -> Float64

Calculate the correlation for a given set of operators applied to specific qubits in either a quantum state vector or a density matrix. This function has two primary forms:

get_corr(state::SparseVector, list_of_operators::String, qubits_applied::Vector): Computes the correlation for a quantum state vector (state) with a specified list of operators and qubits.

get_corr(rho::SparseMatrixCSC, list_of_operators::String, qubits_applied::Vector): Computes the correlation for a density matrix (rho) with a specified list of operators and qubits.

The corr_from_rho function is an alias to get_corr for density matrices.

Arguments

state::SparseVector: The quantum state vector.

rho::SparseMatrixCSC: The density matrix.

list_of_operators::String: A string representing a list of operators, e.g., "Z,Z".

qubits_applied::Vector: A vector of qubit indices on which the operators are applied.

Returns

Float64: The computed correlation value.

Examples

# For a state vector
state = SparseVector([...]) # define your state vector
correlation = get_corr(state, "Z,Z", [2, 4])

# For a density matrix
rho = SparseMatrixCSC([...]) # define your density matrix
correlation = get_corr(rho, "Z,Z", [2, 4])

BlueTangle.get_corr_from_measurement — Function

get_corr_from_measurement(m::Measurement, qubits::Vector) -> Any

Calculate correlation from a measurement on specified qubits.

m: A Measurement object.

qubits: Vector of qubits for measuring correlations, e.g. eg: qubits=[1,3] measure <Z1Z3>

Returns the calculated correlation.

Circuit functions

BlueTangle.compile — Function

Alias:

quantum_circuit(ops::Vector{T}, options::Options=Options()) -> Circuit where T <: QuantumOps
compile(ops::Vector{T}, options::Options=Options()) -> Circuit where T <: QuantumOps
get_circuit_from_ops(ops::Vector{T}, options::Options=Options()) -> Circuit where T <: QuantumOps

Compile a set of quantum operations into a circuit.

ops: A vector of QuantumOps objects.

options: Optional compilation options.

Returns a compiled Circuit object.

BlueTangle.quantum_circuit — Function

quantum_circuit(ops::Vector{T}, options::Options=Options()) -> Circuit where T <: QuantumOps

Alias for compile

BlueTangle.sample — Function

sample(circuit::Circuit, number_of_experiment::Int) -> Measurement

Measure a quantum circuit multiple times.

circuit: A Circuit object.

number_of_experiment: Number of times to execute the circuit.

Returns a Measurement object with the results.

BlueTangle.circuit_to_state — Function

circuit_to_state(circuit::Circuit; init_state::SparseVector=sparse([])) -> SparseVector

Convert a quantum circuit to a state vector.

circuit: A Circuit object.

init_state: (Optional) Initial state vector.

Returns a state vector representing the circuit.

BlueTangle.circuit_to_rho — Function

circuit_to_rho(circuit::Circuit) -> SparseMatrixCSC get_rho_from_circuit get_state_from_circuit

Convert a quantum circuit to a density matrix (rho).

circuit: A Circuit object.

Returns a density matrix representing the circuit.

Plot

BlueTangle.plot_measurement — Function

plot_measurement(sample_probs::Vector; rep::Symbol=:int, basis::String="Z")

Plots the outcome probabilities for quantum measurements.

sample_probs: Vector of tuples, each containing outcome probabilities and number of qubits.

Creates a bar plot showing the probabilities of the most likely outcomes in the specified measurement basis.

plot_measurement(m::Measurement; rep::Symbol=:int)

Plots the outcome probabilities for a single quantum measurement.

m: A Measurement object.

Creates a bar plot showing the probabilities of the most likely outcomes from the measurement.

BlueTangle.plot_circuit — Function

plot_circuit(circuit::Circuit)

Plots a quantum circuit diagram from a Circuit object.

circuit: A Circuit object representing the quantum circuit.

Creates a visual representation of the quantum circuit.

plot_circuit(ops::Vector{QuantumOps}; list_of_initial_qubits::Vector{Int}=Int[])

Plots a quantum circuit diagram from a vector of quantum operations.

ops: Vector of QuantumOps representing the operations in the circuit.

list_of_initial_qubits: (Optional) List of initial qubit states.

Creates a visual representation of the quantum circuit based on the specified operations and initial qubit states.

Quantum Toolkit

BlueTangle.entanglement_entropy — Function

entanglement_entropy(psi::SparseVector) -> Float64

Calculates the entanglement entropy of a quantum state.

psi: The quantum state vector (SparseVector) for which to calculate the entropy.

Returns the calculated entanglement entropy.

entanglement_entropy(rho::SparseMatrixCSC) -> Float64

Calculates the entanglement entropy of a density matrix.

rho: The density matrix (SparseMatrixCSC) for which to calculate the entropy.

Returns the calculated entanglement entropy.

BlueTangle.classical_shadow — Function

classical_shadow(circuit::Circuit, number_of_experiment::Int) -> SparseMatrixCSC

Construct a density matrix (rho) from classical shadow representation.

circuit: A Circuit object.

number_of_experiment: Number of experiments to run.

Returns a sparse density matrix representing the classical shadow.

Other

BlueTangle.get_N — Function

Alias:

get_N(state::SparseVector) -> Int
get_N(rho::SparseMatrixCSC) -> Int

Calculate the number of qubits (N) from a quantum state vector or a density matrix.

state: A sparse quantum state vector.

rho: A sparse density matrix.

Returns N, the number of qubits, calculated as the logarithm base 2 of the length of the state vector or the size of the density matrix.

BlueTangle.expand_multi_op — Function

expand_multi_op(list_of_operators::String, qubits_applied::Vector, N::Int) -> Matrix

Expand multiple quantum operators over a specified set of qubits.

list_of_operators: A string representing a list of operators.

qubits_applied: A vector of qubits on which operators are applied.

N: Total number of qubits.

Returns a matrix representing the expanded operators.

BlueTangle.string_to_matrix — Function

string_to_matrix(list_of_operators::String) -> Matrix

Convert a string of comma-separated operators into a matrix representation.

list_of_operators: A string representing a list of operators, e.g.: "Z,Z,I,P(.2)"

Returns a matrix representation of the operators.

BlueTangle.state_vector_create — Function

state_vector_create(list_of_qubits::Vector) -> SparseVector

Creates a quantum state vector from a list of qubits.

list_of_qubits: A vector representing the state of each qubit.

Returns a sparse vector representing the quantum state of the system.

BlueTangle.apply_op! — Function

apply_op!(state::SparseVector, op::QuantumOps)

Apply a quantum gate operation to a state vector in place.

state: A sparse quantum state vector to be modified.

op: A QuantumOps object representing the gate operation.

Modifies the state vector directly.

BlueTangle.apply_op_rho! — Function

apply_op_rho!(rho::SparseMatrixCSC, op::QuantumOps)

Apply a quantum gate operation to a state vector in place.

rho: A sparse quantum density matrix to be modified.

op: A QuantumOps object representing the gate operation.

Modifies the state vector directly.

Object types

BlueTangle.QuantumOps — Type

QuantumOps

Abstract type representing quantum operations.

BlueTangle.Op — Type

Op(q::Int, name::String, gate::Matrix, qubit::Int, target_qubit::Int, noise::QuantumChannel) <: QuantumOps

Represents a quantum operation.

q: Number of qubits involved in the operation.

name: Name of the operation.

gate: Matrix representation of the quantum operation.

qubit: Index of the target qubit.

target_qubit: Index of the target qubit for two-qubit operations.

noise: Noise model associated with the operation.

Constructs an Op object representing a quantum operation with optional noise.

BlueTangle.ifOp — Type

ifOp(q::Int, name::String, gate::Matrix, qubit::Int, if01::Tuple{Matrix,Matrix}, noise::QuantumChannel) <: QuantumOps

Represents a conditional quantum operation used for mid-circuit born measurements. It is specifically designed for mid-circuit measurements in the X, Y, Z, or a random basis (R). Depending on the measurement outcomes (0 or 1), different gates specified in if01 can be applied.

Fields

q: Integer representing the number of qubits involved in the operation.

name: String indicating the name of the operation. Valid names are "MX", "MY", "MZ", or "MR", corresponding to operations in the X, Y, Z basis, or a random basis (R), respectively.

gate: Matrix representing the quantum operation.

qubit: Integer specifying the index of the target qubit.

if01: Tuple of two matrices. The first matrix is applied if the measured state of the qubit is 0, and the second matrix is applied if the measured state is 1.

noise: Represents a 'born measurement quantum channel', indicating the noise model associated with the operation.

Usage

ifOp is constructed to represent quantum operations that are conditional on the measurement outcome of a qubit. This operator is particularly useful in quantum circuits for implementing dynamic responses based on mid-circuit measurement results.

Examples

```julia

Example of creating an ifOp for a conditional operation based on the X-basis measurement (MX)

conditionalop = ifOp("MX", qubit, (operationif0, operationif_1))

BlueTangle.Measurement — Type

Measurement(int_basis::Union{Vector, UnitRange}, fock_basis::Vector{Vector{Int}}, sample::Vector, expect::Vector, mag_moments::Vector, measurement_basis::String, number_of_experiment::Int, circuit_name::String, number_of_qubits::Int, density_matrix::SparseMatrixCSC)

Represents the result of quantum measurements.

int_basis: Basis of measurement represented as integers or a range.

fock_basis: Fock basis representation of the measurement.

sample: Vector of probabilities.

expect: Expectation values of the measurement.

mag_moments: Magnetic moments obtained from the measurement.

measurement_basis: Measurement basis used.

number_of_experiment: Number of experiments performed.

circuit_name: Name of the circuit used.

number_of_qubits: Number of qubits involved.

density_matrix: Density matrix obtained from the measurement.

Constructs a Measurement object to store the results of quantum measurements.

BlueTangle.QuantumChannel — Type

QuantumChannel(q::Int, model::String, p::Float64)

Represents a quantum noise channel.

q: Number of qubits affected by the noise channel.

model: The noise model type as a string.

p: Probability parameter for the noise model.

kraus: A vector of matrices representing Kraus operators for the channel.

Constructs a QuantumChannel object for specified qubits, noise model, and probability.

BlueTangle.Circuit — Type

Circuit

Represents a quantum circuit.

stats: NamedTuple containing statistics about the circuit.

options: Options object with circuit configurations.

ops: Vector of QuantumOps representing the operations in the circuit.

Constructs a Circuit object representing a quantum circuit.

BlueTangle.Options — Type

Options(circuit_name::String, measurement_basis::String, measurement_error::QuantumChannel, noise1::QuantumChannel, noise2::QuantumChannel, twirl::Bool, swap_error::Bool, density_matrix::Bool)

Represents configuration options for a quantum circuit.

circuit_name: Name of the circuit.

measurement_basis: Measurement basis used in the circuit.

measurement_error: Noise model for measurement error.

noise1: Noise model for single-qubit operations.

noise2: Noise model for two-qubit operations.

twirl: Boolean flag for twirling operations.

swap_error: Boolean flag for swap errors.

density_matrix: Boolean flag to indicate if density matrix should be calculated.

Constructs an Options object with specified settings for a quantum circuit.

Gates

BlueTangle.gate — Constant

gate

A constant that holds common quantum gates and projectors.

This includes:

Identity (I)

Pauli gates (X, Y, Z)

Hadamard (H)

Phase gates (S, T)

Special gates like sqrt(X) (equal to RX(pi/2)exp(1impi/4))

Projectors (proj0 for |0><0|, proj1 for |1><1|)

Controlled gates such as CX (CNOT), CNOT (an alias for CX), and CZ

The SWAP, iSWAP, fSWAP gate

The ECR gate

Each single-qubit gate is represented as a 2x2 matrix, while multi-qubit gates like CNOT, ECR, SYC, CZ, and SWAP, iSWAP, fSWAP are represented as 4x4 matrices.

BlueTangle.gates1 — Function

gates1(op_name::String, param...) -> Matrix

Constructs a matrix representation of a single-qubit gate based on the gate name and parameters.

op_name: Name of the single-qubit gate.

param: Parameters for the gate, if applicable.

Returns a 2x2 matrix representing the specified single-qubit gate.

BlueTangle.gates2 — Function

gates2(op_name::String, param...) -> Matrix

Constructs a matrix representation of a two-qubit gate based on the gate name and parameters.

op_name: Name of the two-qubit gate.

param: Parameters for the gate, if applicable.

Returns a 4x4 matrix representing the specified two-qubit gate.

BlueTangle.U1 — Function

u1(λ), u2(φ, λ), u3(θ, φ, λ),

Functions to construct generic single-qubit rotation gates based on Euler angles.

u1(λ): Rotation about Z-axis.

u2(φ, λ): Rotation about X+Z axis.

u3(θ, φ, λ): Generic single-qubit rotation.

Each function returns a 2x2 matrix representing the specified rotation.

Example for u1(λ): λ=pi -> Z λ=pi/2 -> S λ=pi/4 -> T u1(λ)=U(0, 0, λ)

BlueTangle.U2 — Function

U2(φ, λ)

Single-qubit rotation about the X+Z axis.

(φ, λ)=(0,pi) -> H (φ, λ)=(0,0) -> RY(pi/2) (φ, λ)=(-pi/2,pi/2) -> RX(pi/2)

u2(φ, λ)=U(pi/2, φ, λ)

BlueTangle.U3 — Function

U3(θ, φ, λ)

Generic single-qubit rotation gate with 3 Euler angles.

(θ, φ, λ)=(θ, -pi/2, pi/2)=RX(θ) (θ, φ, λ)=(θ, 0, 0)=RY(θ)

BlueTangle.random_ops — Function

random_ops(N::Int, len::Int; measure_prob::Float64=0.0, measure_basis::Vector{String}=["MX", "MY", "MZ"]) -> Vector{QuantumOps}

Create a sequence of random quantum gate operations, with optional mid-circuit measurements.

Arguments

N::Int: The number of qubits in the system.

len::Int: The length of the sequence of operations to generate.

Keyword Arguments

measure_prob::Float64: The probability of adding a measurement operation after each gate.

measure_basis::Vector{String}: The basis in which measurements are performed.

Returns

Vector{QuantumOps}: A vector of randomly chosen quantum operations (QuantumOps), each representing a gate or a measurement operation.

Description

This function creates a vector of quantum operations, where each operation is either a randomly chosen gate from the set {"X", "Y", "Z", "H", "S", "CX","CZ","CP","GIVENS","FSIM","SWAP","iSWAP","fSWAP","SYC","ECR"} or a measurement operation, based on measure_prob.

Example

ops = random_ops(5, 10; measure_prob=0.2, measure_basis=["MX","MZ"])

This example generates a sequence of 10 random gates and measurements (with a 20% chance of a measurement after each gate) for a 5-qubit system.

BlueTangle.random_clifford — Function

random_clifford(N::Int, len::Int; measure_prob::Float64=0.0, measure_basis::Vector{String}=["MX","MY","MZ"]) -> Vector{QuantumOps}

Generate a random sequence of Clifford gates, with optional mid-circuit measurements.

Arguments

N::Int: The number of qubits in the system.

len::Int: The length of the sequence of operations to generate.

Keyword Arguments

measure_prob::Float64: The probability of adding a measurement operation after each gate.

measure_basis::Vector{String}: The basis in which measurements are performed.

Returns

Vector{QuantumOps}: A vector of randomly chosen quantum operations (QuantumOps), each representing a Clifford gate or a measurement operation.

Description

This function creates a vector of quantum operations, where each operation is either a randomly chosen Clifford gate from the set {"X", "Y", "Z", "H", "S", "CNOT", "SWAP", "CZ"} or a measurement operation, based on measure_prob. For two-qubit gates ("CNOT", "SWAP", "CZ"), adjacent qubits (qubit r and qubit r+1) are selected. For single-qubit gates and measurements, a random qubit r is chosen.

Example

ops = random_clifford(5, 10; measure_prob=0.2, measure_basis=["MX","MZ"])

This example generates a sequence of 10 random Clifford gates and measurements (with a 20% chance of a measurement after each gate) for a 5-qubit system.

Noise

BlueTangle.is_kraus_valid — Function

is_kraus_valid(kraus::Vector{Matrix}) -> Bool

Determines the validity of a set of Kraus operators.

kraus: A vector of matrices, each representing a Kraus operator.

This function checks if the provided Kraus operators form a valid quantum channel. It does so by verifying if the sum of the products of each Kraus operator and its adjoint (approximately) equals the identity matrix. Returns true if the set is valid, false otherwise.

BlueTangle.Noise1 — Function

Noise1(model::String, p::Float64)

Constructors for QuantumChannel objects for 1 qubit.

model: The noise model type as a string.

p: Probability parameter for the noise model.

Returns a QuantumChannel object.

BlueTangle.Noise2 — Function

Noise2(model::String, p::Float64)

Constructors for QuantumChannel objects for 2 qubits.

model: The noise model type as a string.

p: Probability parameter for the noise model.

Returns a QuantumChannel object.

BlueTangle.apply_noise — Function

apply_noise(ops::Vector{T},noise12::Tuple) where T <: QuantumOps -> Vector{QuantumOps}

Applies specified noise models to a vector of quantum operations.

ops: Vector of QuantumOps representing the operations.

noise12: A tuple containing the single-qubit and two-qubit noise models.

Returns a new vector of QuantumOps with noise models applied.

BlueTangle.apply_twirl — Function

apply_twirl(op::QuantumOps, twirling_noise::Union{QuantumChannel,Bool}) -> Vector{QuantumOps}

Applies twirling with specified noise to a single quantum operation.

op: A QuantumOps object representing the quantum operation.

twirling_noise: Noise model to be applied during twirling.

Returns a vector of QuantumOps representing the twirled operation with noise.

BlueTangle.cnot_amplifier! — Function

cnot_amplifier!(ops::Vector{T}, CNOT_pair=0) where T <: QuantumOps

Amplify the presence of CNOT (or CX) operations in a vector of quantum operations.

This function adds extra pair of CNOT operation in the ops vector a specific number of times in place. This is useful for amplifying the effect of noise via CNOT operations in a sequence of quantum operations.

Arguments

ops::Vector{T}: A vector of quantum operations, where T is a subtype of QuantumOps.

CNOT_pair::Int (optional): The number of additional pairs of CNOT operations to insert for each original CNOT operation in ops. The default value is 0, which means no additional operations are inserted.

Examples

ops = [Op("H",1), Op("CNOT",1,2), Op("X",2)]
cnot_amplifier!(ops, 1)
# `ops` will be modified to: [Op("H",1), Op("CNOT",1,2), Op("CNOT",1,2), Op("CNOT",1,2), Op("X",2)]

BlueTangle.op_amplifier! — Function

op_amplifier!(ops::Vector{T},op_pair=0) where T <: QuantumOps

same as cnot_amplifier! but for all operations

BlueTangle.error_mitigate_data — Function

error_mitigate_data(xdata::Vector, ydata::Vector)

Perform error mitigation on a dataset by fitting a linear model and extracting the estimate and standard error.

This function takes two vectors xdata and ydata which represent the independent and dependent variables of a dataset, respectively.

Arguments

xdata::Vector: The independent variable data points.

ydata::Vector: The dependent variable data points, corresponding to each xdata point.

Returns

est: The estimated intercept from the linear fit.

se: The standard error of the estimated intercept.

fit_plot: A tuple containing the x-values from 0 to the last element of xdata and the corresponding fitted y-values from the model.

Hamiltonian

BlueTangle.trotter_ising — Function

trotter_ising(t::Float64, N::Int, J::Float64, h::Float64; dt=0.1) -> Vector{QuantumOps}

Simulates the Ising model Hamiltonian using Trotterization.

t: The total time for the simulation.

N: Number of qubits in the Ising chain.

J: Coupling constant for the interaction between adjacent qubits.

h: Magnetic field strength applied in the x-direction.

dt (optional): Time step for the Trotterization. Default is 0.1.

Performs Trotterization of the Ising model Hamiltonian over N qubits for a total time t, given the coupling constant J and magnetic field strength h. The function constructs a series of quantum gates that approximate the evolution under the Ising Hamiltonian.

Returns a vector of QuantumOps, representing the sequence of operations for the Trotterized Ising model simulation.

Index

BlueTangle.gate

BlueTangle.Circuit

BlueTangle.Measurement

BlueTangle.Op

BlueTangle.Options

BlueTangle.QuantumChannel

BlueTangle.QuantumOps

BlueTangle.ifOp

BlueTangle.Noise1

BlueTangle.Noise2

BlueTangle.U1

BlueTangle.U2

BlueTangle.U3

BlueTangle.apply_noise

BlueTangle.apply_op!

BlueTangle.apply_op_rho!

BlueTangle.apply_twirl

BlueTangle.circuit_to_rho

BlueTangle.circuit_to_state

BlueTangle.classical_shadow

BlueTangle.cnot_amplifier!

BlueTangle.compile

BlueTangle.entanglement_entropy

BlueTangle.error_mitigate_data

BlueTangle.expand_multi_op

BlueTangle.gates1

BlueTangle.gates2

BlueTangle.get_N

BlueTangle.get_corr

BlueTangle.get_corr_from_measurement

BlueTangle.get_expect

BlueTangle.get_expect_from_measurement

BlueTangle.get_probabilities_from_sample

BlueTangle.is_kraus_valid

BlueTangle.op_amplifier!

BlueTangle.plot_circuit

BlueTangle.plot_measurement

BlueTangle.quantum_circuit

BlueTangle.random_clifford

BlueTangle.random_ops

BlueTangle.sample

BlueTangle.sample_outcomes

BlueTangle.state_vector_create

BlueTangle.string_to_matrix

BlueTangle.trotter_ising